requirements.txt
Django>=4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
dj-rest-auth==4.0.1
django-allauth==0.54.0
django-environ>=0.4.5
django-extensions==3.1.3
django-stubs==1.8.0
gunicorn==20.1.0
tzdata==2023.3
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10